// types.ts
export type View = 'Dashboard' | 'Products' | 'Billing' | 'Analytics';
export type Theme = 'light' | 'dark';

export interface Product {
  sku: string;
  user_id: string;
  name: string;
  costPrice: number;
  price: number;
  quantity: number;
  lowStockThreshold: number;
}

export interface CartItem {
  productSku: string;
  name: string;
  price: number;
  quantity: number;
}

// Represents an item within a completed transaction
export interface TransactionItem {
  productSku: string;
  name: string; // Denormalized for easier receipt rendering
  quantitySold: number;
  pricePerItem: number; // Price at the time of sale
}

export interface Transaction {
  id?: number; // Optional because it's generated by the database
  user_id: string;
  timestamp: Date;
  items: TransactionItem[];
  subtotal: number;
  discountPercent: number;
  total: number;
}